+++ actionview/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Remove deprecated block passing to FormBuilder#new.
+
+    *Vipul A M*
+
 *   Pick `DateField` `DateTimeField` and `ColorField` values from stringified options allowing use of symbol keys with helpers.
 
     *Jon Rowe*